remove authz folder
