+++ src/seacat-auth/components/ButtonWithAuthz.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
-import { default as ControlledSwitch } from '../../../components/ControlledSwitch';
-import { authz } from '../../utils/authz';
+import { Button } from 'reactstrap';
+import { authz } from '../utils/authz';
 
 /*
-	ControlledSwitchWithAuthz component emulates the functionality of asab-webui <ControlledSwitch/> component
+	ButtonWithAutz component emulates the functionality of reactstrap <Button/> component
 	and extend it by disable feature based on resources.
 
 	* resource
@@ -16,7 +16,7 @@ import { authz } from '../../utils/authz';
 		* the option to hide the button completely instead of disabling it, default is false
 
 	* Language localisations for generic `unauthorized message` can be added to the translation.json files of
-	  public/locales/en & public/locales/cs of the product where component ControlledSwitchWithAuthz is used. The default
+	  public/locales/en & public/locales/cs of the product where component ButtonWithAuthz is used. The default
 	  message is `You do not have rights` and it can be re-set in locales as e.g.
 
 	  {
@@ -27,7 +27,7 @@ import { authz } from '../../utils/authz';
 
 	Example:
 
-	import { ControlledSwitchWithAuthz } from 'asab_webui_components';
+	import { ButtonWithAuthz } from 'asab_webui_components';
 	import { useSelector } from 'react-redux';
 
 	...
@@ -38,21 +38,24 @@ import { authz } from '../../utils/authz';
 
 	return(
 		...
-			<ControlledSwitchWithAuthz
-				title={t(`MyModule|${isDisabled ? "Enable" : "Disable"} file`)}
+			<ButtonWithAuthz
+				title={t('MyApp|Delete')}
+				color="danger"
 				size="sm"
-				isOn={!isFileDisabled}
-				toggle={switchFileState}
+				onClick={() => { deleteApp() }}
 				resource="myapp:mylist:write"
 				resources={resources}
 				hideOnUnauthorizedAccess={true}
-			/>
+				>
+				<span className="bi bi-code-slash pe-2" />
+				{t('MyApp|Delete')}
+			</ButtonWithAuthz>
 		...
 	)
 
 */
 
-export function ControlledSwitchWithAuthz(props) {
+export function ButtonWithAuthz(props) {
 	let childProps = {...props}; // Create a new child component to have option to remove props
 	let authzObj = authz(childProps);
 
@@ -62,10 +65,12 @@ export function ControlledSwitchWithAuthz(props) {
 
 	return (
 		hide && disabled ? null :
-		<ControlledSwitch
+		<Button
 			{...childProps}
 			title={title}
 			disabled={disabled}
-		/>
+		>
+			{childProps.children}
+		</Button>
 	)
 }
