+++ src/seacat-auth/components/ControlledSwitchWithAuthz.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
-import { default as UncontrolledSwitch } from '../../../components/UncontrolledSwitch';
-import { authz } from '../../utils/authz';
+import { default as ControlledSwitch } from '../../components/ControlledSwitch';
+import { authz } from '../utils/authz';
 
 /*
-	UncontrolledSwitchWithAuthz component emulates the functionality of asab-webui <UncontrolledSwitch/> component
+	ControlledSwitchWithAuthz component emulates the functionality of asab-webui <ControlledSwitch/> component
 	and extend it by disable feature based on resources.
 
 	* resource
@@ -16,7 +16,7 @@ import { authz } from '../../utils/authz';
 		* the option to hide the button completely instead of disabling it, default is false
 
 	* Language localisations for generic `unauthorized message` can be added to the translation.json files of
-	  public/locales/en & public/locales/cs of the product where component UncontrolledSwitchWithAuthz is used. The default
+	  public/locales/en & public/locales/cs of the product where component ControlledSwitchWithAuthz is used. The default
 	  message is `You do not have rights` and it can be re-set in locales as e.g.
 
 	  {
@@ -27,7 +27,7 @@ import { authz } from '../../utils/authz';
 
 	Example:
 
-	import { UncontrolledSwitchWithAuthz } from 'asab_webui_components';
+	import { ControlledSwitchWithAuthz } from 'asab_webui_components';
 	import { useSelector } from 'react-redux';
 
 	...
@@ -38,12 +38,11 @@ import { authz } from '../../utils/authz';
 
 	return(
 		...
-			<UncontrolledSwitchWithAuthz
+			<ControlledSwitchWithAuthz
 				title={t(`MyModule|${isDisabled ? "Enable" : "Disable"} file`)}
 				size="sm"
-				register={register}
-				setValue={setValue}
-				name='setTotp'
+				isOn={!isFileDisabled}
+				toggle={switchFileState}
 				resource="myapp:mylist:write"
 				resources={resources}
 				hideOnUnauthorizedAccess={true}
@@ -53,7 +52,7 @@ import { authz } from '../../utils/authz';
 
 */
 
-export function UncontrolledSwitchWithAuthz(props) {
+export function ControlledSwitchWithAuthz(props) {
 	let childProps = {...props}; // Create a new child component to have option to remove props
 	let authzObj = authz(childProps);
 
@@ -63,7 +62,7 @@ export function UncontrolledSwitchWithAuthz(props) {
 
 	return (
 		hide && disabled ? null :
-		<UncontrolledSwitch
+		<ControlledSwitch
 			{...childProps}
 			title={title}
 			disabled={disabled}
